# apps/inngest-app/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inngest App API",
    "version": "1.0.0",
    "description": "API documentation for the Inngest app"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your bearer token in the format: Bearer {token}"
      }
    },
    "schemas": {
...
      "WorkflowInputField": {
        "type": "object",
        "required": [
          "key",
          "label",
          "type"
        ],
        "properties": {
          "key": {
            "type": "string",
            "minLength": 1
          },
          "label": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "number",
              "boolean",
              "json",
              "credential",
              "email",
              "url",
              "date",
              "time",
              "file",
              "select",
              "multiselect"
            ]
          },
          "required": {
            "type": "boolean",
            "default": false
          },
          "defaultValue": {},
          "validation": {
            "type": "object",
            "properties": {
              "min": {
                "type": "number"
              },
              "max": {
                "type": "number"
              },
              "pattern": {
                "type": "string"
              },
              "options": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "WorkflowTemplateResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "category",
          "eventName"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "REPORTING",
              "DATA_PROCESSING",
              "COMMUNICATION",
              "AUTOMATION",
              "ANALYTICS",
              "INTEGRATION",
              "MONITORING",
              "CUSTOM"
            ]
          },
          "version": {
            "type": "string",
            "default": "1.0.0"
          },
          "eventName": {
            "type": "string"
          },
          "canBeScheduled": {
            "type": "boolean",
            "default": false
          },
          "requiredProviders": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "GOOGLE",
                "SLACK",
                "HUBSPOT",
                "FIRECRAWL",
                "CUSTOM"
              ]
            }
          },
          "inputFields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowInputField"
            }
          },
          "outputSchema": {
            "type": "object"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "WorkflowResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "enum": [
              "REPORTING",
              "DATA_PROCESSING",
              "COMMUNICATION",
              "AUTOMATION",
              "ANALYTICS",
              "INTEGRATION",
              "MONITORING",
              "CUSTOM"
            ]
          },
          "workflowDefinitionId": {
            "type": "string"
          },
          "available": {
            "type": "boolean",
            "description": "Whether the workflow instance is available for use by the user"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the workflow is currently active/scheduled to run"
          },
          "canBeScheduled": {
            "type": "boolean"
          },
          "cronExpressions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timezone": {
            "type": "string",
            "nullable": true
          },
          "lastRunAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "nextRunAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "input": {
            "type": "object",
            "nullable": true
          },
          "eventName": {
            "type": "string"
          },
          "requiredProviders": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "GOOGLE",
                "SLACK",
                "HUBSPOT",
                "FIRECRAWL",
                "CUSTOM"
              ]
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "credentials": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                },
                "provider": {
                  "type": "string"
                },
                "config": {
                  "type": "object",
                  "nullable": true
                }
              }
            }
          },
          "template": {
            "$ref": "#/components/schemas/WorkflowTemplateResponse"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/workflows/{id}": {
      "get": {
        "summary": "Get a workflow",
        "description": "Returns a single workflow for the authenticated user.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single workflow."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Workflow not found."
          },
          "500": {
            "description": "Error getting workflow."
          }
        }
      },
      "put": {
        "summary": "Update a workflow",
        "description": "Updates a single workflow for the authenticated user. Validates input against the workflow template schema.",
        "tags": [
          "Workflows"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The workflow ID to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Custom name for the workflow"
                  },
                  "description": {
                    "type": "string",
                    "description": "Workflow description"
                  },
                  "available": {
                    "type": "boolean",
                    "description": "Whether the workflow is available"
                  },
                  "input": {
                    "type": "object",
                    "description": "Workflow input parameters that will be validated against template schema"
                  },
                  "credentials": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "credentialId": {
                          "type": "string"
                        }
                      }
                    },
                    "description": "Credential IDs to associate with this workflow"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workflow updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "400": {
            "description": "Input validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Workflow not found"
          },
          "500": {
            "description": "Error updating workflow"
          }
        }
      },
      "delete": {
        "summary": "Delete a workflow",
        "description": "Deletes a single workflow for the authenticated user.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow deleted successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Error deleting workflow."
          }
        }
      }
    },
...
    "/api/workflows/install/{templateId}": {
      "post": {
        "summary": "Install a workflow template",
        "description": "Creates a new workflow instance from a marketplace template for the authenticated user.",
        "tags": [
          "Workflows"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "templateId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The workflow template ID to install"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Custom name for this workflow instance"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workflow template installed successfully"
          },
          "400": {
            "description": "Invalid template or validation failed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Template not found or user not found"
          },
          "409": {
            "description": "Template already installed"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}